refactored JWS related utilities
