+++ src/core/message.ts
@@ -5,7 +5,7 @@ import { CID } from 'multiformats/cid';
 import { computeCid } from '../utils/cid.js';
 import { GeneralJws } from '../jose/jws/general/types.js';
 import { GeneralJwsSigner } from '../jose/jws/general/signer.js';
-import { GeneralJwsVerifier } from '../jose/jws/general/verifier.js';
+import { Jws } from '../utils/jws.js';
 import { lexicographicalCompare } from '../utils/string.js';
 import { RecordsWriteMessage } from '../interfaces/records/types.js';
 import { validateJsonSchema } from '../validator.js';
@@ -35,7 +35,7 @@ export abstract class Message {
 
   constructor(message: BaseMessage) {
     this.message = message;
-    this.authorizationPayload = GeneralJwsVerifier.decodePlainObjectPayload(message.authorization);
+    this.authorizationPayload = Jws.decodePlainObjectPayload(message.authorization);
 
     this.author = Message.getAuthor(message);
   }
@@ -64,7 +64,7 @@ export abstract class Message {
    * Gets the DID of the author of the given message.
    */
   public static getAuthor(message: BaseMessage): string {
-    const author = GeneralJwsVerifier.getSignerDid(message.authorization.signatures[0]);
+    const author = Jws.getSignerDid(message.authorization.signatures[0]);
     return author;
   }
